Ready to go
Much work later this script seems to work neatly to test the small use case we want to test for applications to become a supplier
