+++ conf/logback.xml
@@ -12,7 +12,7 @@
 	<!-- 	<logger name="io.gatling.http.ahc" level="TRACE" /> -->
 	<!--    <logger name="io.gatling.http.response" level="TRACE" /> -->
 	<!-- Uncomment for logging ONLY FAILED HTTP request and responses -->
-	<!-- 	<logger name="io.gatling.http.ahc" level="DEBUG" /> -->
+	 	<logger name="io.gatling.http.ahc" level="DEBUG" />
 	<!--    <logger name="io.gatling.http.response" level="DEBUG" /> -->
 
 	<root level="WARN">
